01. Base design and REST API Congratulations! You finished the first class of our course. And what do we learn?
Flutter Themes: We learned that when we are going to continue a project that has already started, we often need to do some basic configurations so that, in the end, we can deliver what was intended. In our case, the appearance of our application was completely different from the prototype order and that is why we made the change using the ThemeData.
API and REST: We saw that, to talk about WebAPI, we need to have at least one introduction to what API and REST are. We understand that API is an interface that links communication between client and server, and that REST is a set of rules that govern the implementation and operation of the APIs that are RESTful.