03. Registering with the API
Another class finished and how about we review what we learn?

Asynchronism: We started by doing a brief review of asynchronism in Dart, as we noticed that when we are dealing with requests for Web APIs, these operations will not be instantaneous, that is, they will take time to happen, therefore, they will be asynchronous.
Input Addition Screen: We also saw that it was necessary to build a screen responsible for capturing from the user person the information that will be sent to the Web API.
POST Method: Next, we use the POST method to send the captured information to our Web API. For that, it was necessary to convert our object to a JSON, configure the header to let us know that we are sending the information using this pattern and check if the server response indicates that the entry was created successfully.
SnackBar: Finally, we note that it is extremely important to give feedback to the user regarding the success of the asynchronous operation. For this, we use a SnackBar that will be shown on the screen when the server response is a 201 Created.