04. Listing on the Home Screen
Uhuu!! We arrived at our last class of the course. Shall we remember what we learned?

Searching for the API: After learning how to add new information to a resource using POST, the time had come for us to learn how to read the resource entries. For this we use the GET method to access the journals from our server and thus receive a list containing all registered entries.
Listing on the Screen: With the list of entries available via our operation with GET, we just needed to be able to show on the main screen. For this, we converted back the JSON that arrives as an answer to a list of Journals, and with that list we can replace the database random that we used for that real information received via Web API.
Refactoring the code: Finally, we refactored our code so that it was much more structured and readable.