01. Base Design and HTTP methods
Congratulations! You finished the first class of our course. And what do we learn?

Changing with PUT: We saw that, since we were able to add and read entries to our resource via HTTP, the next natural step is to allow users to change those entries. For this, we use the PUT method in HTTP, passing the input ID to be changed in the URL and the entire body of the object Journal like JSON no body of the request.
Removing with DELETE: Finally, it was necessary to remove an entry. We saw that, in fact, this is one of the simplest operations for our case. It was enough to use the HTTP DELETE method, and the input ID is changed at the URL.
Confirmation Dialog: Because removal is a destructive operation, we noticed that it would be interesting to add a layer of protection against accidental cases. For this we created a generic confirmation dialog ( generic, as it can be used in other situations ), and only when the user confirms the desire to perform this operation will we remove the entry.