02. Authorization flow
Class 2 successfully completed? So, let's remember what we learned:

Installing JSON Server Auth: We saw that, as it was, our server was not prepared to handle authentication of user people. To enable this functionality we installed JSON Server Auth.
Understanding the Authorization Flow: We talked about that each API will have specifics in the way of authenticating and authorizing users, but that the system that uses User - Password - Token, or just Authentication with Token, it is very common and it is what we will use with JSON Server Auth.
Postman: We know Postman, a tool for executing and detailing requests and responses. It is very useful for doing testing operations quickly for our API, freely changing the information sent ( as headers and body ) and carefully analyzing the response received.