04. Flutter Authentication
Another class finished and how about we review what we learn?

Importing Login Screen: We saw that we will not always need to create a tool from scratch, often part of the code that we are going to use is ready and just import it into our code. This was exactly the case with our Login Screen, we received the code in Dart ready and we just need to import it into our file directory and make small settings.
Login via API: Right after that, we needed to send the email and password information to the API, so that the server could try to authenticate with this information. For this, we capture this data on the screen we import and create a Service ( service ) called auth_service. We did this to separate requests for Journal referring to authentication. Then just send the email and password information via POST, and receive the response from the server.
Registering via API: We noticed that, in order not to need to create a registration screen, we could suggest registering a new user if the email passed in the Login was invalid. To do this test, we checked the Status Code of the login response and, in case of failure, we checked if the body of the response contained the text that we saw as the error message for cases where the server did not find the email passed in the request in its database. In such cases, we open a confirmation Dialog suggesting the user to create a new account with this email and password.
Saving Locally: Finally, we know that soon afterwards we will use the information from Token and the user ID to make our authenticated requests. So we've already done the job of persistently saving, using Shared Preferences, that information in the device's memory.