04. Authorizing Requests
Another class finished and how about we review what we learn?

Initial Screen Redirection: We realized that if the user has already logged in once, he will already have the authentication token, so it makes no sense to send him to the login screen again. To solve this we made a check if the token is null already in our main, if so, we make our initialRoute direct to login. If the token is different from null, our initialRoute receives the value home.
Authentication Header: We learned that we need to authenticate our already built operations from GET, POST, PUT and DELETE. For this, we understand in Postman how it works to add an authentication header with Token in an HTTP request, and we obtained and added that information to all operations in JournalService.
Displacing the User: Finally, we learned that it is necessary to give the user the option to log out of their account. For that, we created a Drawer with an exit button. When clicked, this button clears the information stored in Shared Preferences and calls the Login Screen.