05. Listing Home Screen
Uhuu!! We arrived at our last class of the course. Shall we remember what we learned?

Preparing Exceptions Services: We saw that when we are dealing with communication with Web APIs, a lot can go wrong or not go as we expected. Therefore, it is necessary to deal with these problems in order to: either offer a solution to the user, or inform them of what happened in the most pleasant way possible. For this, we have created some personalized exceptions for specific cases in our services that we can handle, and we use the possibility to send a message using the HTTPException for generalist situations.
Treating Exceptions: We learned that now that we are making exceptions to our services, it is necessary to revisit each place where we use a method of these services and use the method catchError of Future to notice whether or not our operation raised a problem, and to treat it.
Adding a Timeout: We realized that our requests cannot wait for a response from the server indefinitely and to resolve this, we added a Timeout to our HTTP Client, and a TimeoutException will be launched if the time we set expires. We take advantage of the structure of catchError that we created earlier to deal with this case as well.
Refactoring the code: Finally, we refactored our code so that it was much more structured and readable.